Adds Buttons, Switches, and Doors
Adds button logic and switches to the game.  Logic can be customized so multiple buttons can be tied to the same door, and vice versa.  Included is a Logic Testing Scene to, well, test logic.  Have fun and let me know in Discord what you think or if you have any ideas for other puzzle logic.

-Jared